+++ tools/bulk-publish/index.html
@@ -1,0 +1,19 @@
+<html lang="en">
+
+<head>
+    <title>Bulk Publish</title>
+    <link rel="stylesheet" href="../../libs/styles/variables.css">
+    <link rel="stylesheet" href="../../libs/styles/styles.css">
+    <link rel="stylesheet" href="./css/bulk.css">
+</head>
+
+<body>
+<div class="bulk"></div>
+<script type="module">
+  import init from './js/bulk.js';
+  const bulk = document.querySelector('.bulk');
+  init(bulk);
+</script>
+</body>
+
+</html>
